feat : charcustomization 최종/mock data 사용
