+++ pages/side-panel/package.json
@@ -32,8 +32,10 @@
     "@observablehq/plot": "^0.6.17",
     "@types/d3": "^7.4.3",
     "d3": "^7.9.0",
+    "dom-to-image": "^2.6.0",
     "html2canvas": "^1.4.1",
-    "lucide-react": "^0.511.0"
+    "lucide-react": "^0.511.0",
+    "recharts": "^2.15.3"
   },
   "devDependencies": {
     "@extension/tailwindcss-config": "workspace:*",
