+++ pages/side-panel/src/features/FinChartCustom.tsx
@@ -1,0 +1,1 @@
+import { ChartCustom } from './chart-customization/ChartCustom';
