--- pages/side-panel/src/features/chart-customization/README.md
@@ -1,69 +1,0 @@
-# Chart Customization Feature
-
-## 개요
-이미지 형태의 차트를 데이터로 변환하고 다양한 형태로 커스터마이징할 수 있는 기능입니다.
-
-## 담당자
-[담당자 이름]
-
-## 주요 기능
-- 이미지 차트 데이터 추출
-- 차트 타입 변경 (막대, 선, 파이 등)
-- 색상 및 스타일 커스터마이징
-- 차트 내보내기
-
-## 개발 가이드
-
-### 1. 컴포넌트 구조 (참고)
-```
-chart-customization/ 
-├── components/
-│ ├── ChartEditor.tsx # 차트 편집 UI
-│ ├── ColorPicker.tsx # 색상 선택
-│ └── ChartTypeSelector.tsx
-├── hooks/
-│ └── useChartCustomization.ts
-└── types.ts
-```
-
-
-### 2. 주요 타입 정의
-```typescript
-interface ChartData {
-  type: string;
-  data: any;
-  options: {
-    colors?: string[];
-    labels?: string[];
-  };
-}
-```
-
-### 3. 개발 순서
-1. 이미지 클릭 이벤트 처리
-2. 서버 API 연동
-3. 차트 편집 UI 구현
-4. 내보내기 기능 구현
-
-### 4. API 통신
-- 서버 API 호출
-  ```typescript
-  const processChart = async (imageData: string) => {
-    const response = await fetch('/api/chart/process', {
-      method: 'POST',
-      body: JSON.stringify({ image: imageData })
-    });
-    return response.json();
-  };
-  ```
-
-### 5. 스타일링 가이드
-- 차트 라이브러리 통합
-- 커스텀 컨트롤 스타일링
-- 반응형 디자인
-
-### 6. 테스트 체크리스트
-- [ ] 이미지 처리
-- [ ] 차트 변환
-- [ ] 커스터마이징 기능
-- [ ] 내보내기 기능
