+++ pages/side-panel/src/index.tsx
@@ -1,16 +1,12 @@
 import { createRoot } from 'react-dom/client';
-import { ChartCustom } from '@src/features/chart-customization/ChartCustom';
 import { mockCharts } from '@extension/shared';
-import ChartColorCustom from './features/chart-customization/ChartColorCustom';
+import ChartColorCustom from './features/chart-color-customization/ChartColorCustom';
 import '@src/index.css'; // Tailwind 등 포함된 스타일
-import ChartCustomizationType from '@src/features/chart-customization-type/ChartPreview';
 import { Divider } from '@mui/material';
 
 function App() {
   return (
     <>
-      <ChartCustom chart={mockCharts[0]} />
-      <Divider />
       <ChartColorCustom />
     </>
   );
